# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel  Repository Service Pattern ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/employees": {
            "get": {
                "tags": [
                    "employees"
                ],
                "summary": "list",
                "description": "list of the employees",
                "operationId": "employeeList",
                "parameters": [
                    {
                        "name": "full_name",
                        "in": "query",
                        "description": "full_name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "address",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "phone_number",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "+14"
                        }
                    },
                    {
                        "name": "position_id",
                        "in": "query",
                        "description": "position_id",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "department_id",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "is active",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "age_gte",
                        "in": "query",
                        "description": "age greater than or equal",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "age_lte",
                        "in": "query",
                        "description": "age less than or equal",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": "10"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": "10"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": "20"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Token is Expired"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/employees/{id}": {
            "get": {
                "tags": [
                    "employees"
                ],
                "summary": "show",
                "description": "get employee by id",
                "operationId": "employeeShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "full_name": {
                                            "type": "string"
                                        },
                                        "birth_date": {
                                            "type": "string"
                                        },
                                        "is_active": {
                                            "type": "boolean"
                                        },
                                        "position_id": {
                                            "type": "integer"
                                        },
                                        "department_id": {
                                            "type": "integer"
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "age": {
                                            "type": "integer"
                                        },
                                        "phone_number": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        },
                                        "position": {
                                            "$ref": "#/components/schemas/Position"
                                        },
                                        "department": {
                                            "$ref": "#/components/schemas/Department"
                                        },
                                        "courses": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Token is Expired"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/employees/create": {
            "post": {
                "tags": [
                    "employees"
                ],
                "summary": "create",
                "description": "Create employee",
                "operationId": "employeeCreate",
                "requestBody": {
                    "description": "Pass employee data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "birth_date",
                                    "position_id",
                                    "department_id",
                                    "address",
                                    "phone_number"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Ivan Ivanov"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "1998-05-05"
                                    },
                                    "position_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Almaty, Abay st. 45"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+77071234578"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong request input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Token is Expired"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/employees/destroy/{id}": {
            "delete": {
                "tags": [
                    "employees"
                ],
                "summary": "destroy",
                "description": "Destroy employee",
                "operationId": "employeeDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Token is Expired"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/employees/{id}/courses/detach": {
            "delete": {
                "tags": [
                    "employees"
                ],
                "summary": "course detach",
                "description": "Detach course from the employee",
                "operationId": "employeeCourseDetach",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Token is Expired"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Position": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "label": {
                        "description": "Position label",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "slug": {
                        "description": "Position slug",
                        "type": "string",
                        "readOnly": "true",
                        "example": "dev"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Position"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "in": "header",
                "bearerAuth": "bearerAuth",
                "bearerFormat": "JWT"
            }
        }
    }
}